R 4.2.1 tmc-langs-rust Dockerfile
This makes the Dockerfile to build and installs R 4.2.1 properly on the
main Dockerfile (based on Debian Stretch). The size of the full image is

    tmc-langs-cli   0.10.0.4                   4.41GB

and it has both R 3.6.3 and R 4.2.1 installed (the R 4.2.0 version was
4.37GB). The way to get the size down is to make a multistage build and
then replace development libraries with the corresponding runtime
libraries (the effect of this should be approximately 200-300 MB).

Other parts of the Dockerfile provide easier and more size reduction
possibilities.

The keys required for R were moved to their own place, to make the
meaning of them for the main Dockerfile clearer.

This is based on installer for R 4.2.0 on Debian Stretch (omitted) and
only the installer scripts need to be updated. It builds cleanly unless
there are transient network issues (if a download fails, the building fails).