R 3.6.3 on stretch: fix for backport blocking bug
This fixes a bug introduced in the commit 8eed5aa482551ea34bb51dac5d6797e43f361e87
from 22.6.2020.

A misplaced double quote prevented the backport of R 3.6.3 to be
used. This did not block the build until 30.11.2020 which is the time a
required `processx` package was deprecated and removed from R 3.3.3.

The usual propagation of the dep the 2-3 weeks since this change has
propagated throughout the R mirrors, meaning that the actual date for
this had occured is between 14.12.-21.12.2020 before it started
manifesting itself. It seems that the December 15th 2020 version is
still in use, so it must have occured between 16.-21. on December 2020.

In addition commit makes the backport work after a change in November 2021.