Separate R 4.2.1 version on stretch
This is based on older R 4.2.0 version, with only minimal changes
needed. That was based on R 3.6.3 on stretch.

This makes the Dockerfile to build and install R 4.2.1 (and only that).
The size of the image is

    tmc-langs-cli   0.10.0.4_testmain_r421     1.42GB

and it still has both R 3.6.3 and R 4.2.1 installed. Cleaning the build
directory reduced the size by 300MB. The size on R 3.6.3 version (that
is omitted) would be

    tmc-langs-cli   0.10.0.4_testmain_r363     1.03GB

so there is not much to reduce more. The way to get the size down is to
make a multistage build and then replace development libraries with the
corresponding runtime libraries (the effect of this should be could be
300-400 MB).