Update R image builder script to support R-4.4.x
The previous mirror repository is giving packages that are too old for
R-4.4.x to work so it was changed to the main mirror. The current
R-4.4.1 (Race for Your Life) was released in 14.6.2024.

Moreover, the installer recognises the Debian version used in the host
server (currently Buster (oldoldstable), Bullseye (oldstable) and
Bookworm (stable)). The next would be Trixie (testing) which will be
handled in a later commit.